NetBSD/arm: disable pax mprotect restrictions for emu
ARM has separate read and execute protection bits so after writing out
JIT code we need to make it executable with mprotect(2), but PAX
wouldn't let us by default.  Mark the emu binary so that this
restriction is lifted.

On other machines where this is not an issue set PAXCTL to a no-op.