comp-arm.c: save callee-saved registers on enter
In the preamble code (comvec) save callee-saved registers clobbered by
the JIT code and arrange to restore them on return.  Makes JIT work on
arm when xec.c compiled with optimizations.

r11 doesn't seem to be used by the JIT, but we need to keep 8 byte
alignment of the stack pointer.