comp-sparc.c: set patch[size]
punt with WRTPC looks at where the current instruction _ends_ and
that's where the next instruction starts, i.e. patch[pc+1], so if
pc=size (e.g. "raise" as the last instruction in a module) punt needs
to look at patch[size].  Previously it would read garbage past the end
of the array.